+++ CHANGES.md
@@ -9,6 +9,9 @@ Version History
 - Add ContainerState#getHost as a replacement for getContainerIpAddress (see also https://github.com/testcontainers/testcontainers-java/pull/2742)
 - Added additional url params in JdbcDatabaseContainer (see also https://github.com/testcontainers/testcontainers-java/issues/1802)
 - For compatibility with Jaybird 4, when legacy client auth is enabled and `authPlugins` URL param has not been explicitly added, add URL param `authPlugins` with value `Srp256,Srp,Legacy_auth`
+- Updated default image version to Firebird 3.0.5.
+
+  Because of intermittent connection problems in Firebird 3.0.6 (CORE-6346, CORE-6347, CORE-6348), 3.0.6 is not used as the default. Firebird 3.0.7 will fix this issue. 
 
 1.0.3
 -----